moved get foods code to mariadb service
