+++ flask/api/foods/__init__.py
@@ -1,49 +1,16 @@
 import os
 from flask import Flask, request, jsonify
 from flask_restful import Resource, Api
-import mysql.connector as mariadb
-import sys
-from database import load_csv
+from services import MariaDBService
 
 # Foods
 # shows a list of all foods, and lets you POST to add a new food
 class Foods(Resource):
-
-  # Initialize by establishing connection to mariadb
-    #def __init__(self):
-    #    load_csv()
+  def __init__(self):
+    self.db = MariaDBService()
 
   #  Get all foods
   #  Used to populate app drop-down
-    def get(self):
-        print("getting all foods...")
-    
-    # Connect to MariaDB
-    # this function needs to be stored somewhere else but i couldnt get import to work
-        try:
-            conn = mariadb.connect(
-                user=os.environ.get("MYSQL_USER", "admin"),
-                password=os.environ["MYSQL_PWD"],
-                host=os.environ.get("MYSQL_HOST", "mariadb"),
-                port=3306,
-                database="kaggledb")
-            
-            # Get cursor and query MariaDB - eager loading
-            cur = conn.cursor(buffered=True)
-            cur.execute("Select database();")
-            cur.execute("SELECT food_product FROM food_production;")
-
-            # Add each row to list
-            foods = []
-            for food_product in cur:
-                foods.append(food_product[0])
-            
-            # return results
-            return foods
-
-        # Throw error msg and exit
-        except mariadb.Error as e:
-            print(f"Error while connection to Mariadb: {e}")
-            sys.exit(1)
-
-        return "error"
+  def get(self):
+    print("getting all foods...")
+    return self.db.get_all_foods()
