--- flask/api/services/emissions_service.py
@@ -1,5 +1,0 @@
-from database import load_csv
-
-class EmissionsService(self):
-  def __init__(self):
-    self.
